bashrs::bash_parser::ast

Enum BashExpr

pub enum BashExpr {

    Literal(String),
    Variable(String),
    CommandSubst(Box<BashStmt>),
    Arithmetic(Box<ArithExpr>),
    Array(Vec<BashExpr>),
    Concat(Vec<BashExpr>),
    Test(Box<TestExpr>),
    Glob(String),
    DefaultValue {
        variable: String,
        default: Box<BashExpr>,
    },
    AssignDefault {
        variable: String,
        default: Box<BashExpr>,
    },
    ErrorIfUnset {
        variable: String,
        message: Box<BashExpr>,
    },
    AlternativeValue {
        variable: String,
        alternative: Box<BashExpr>,
    },
    StringLength {
        variable: String,
    },
    RemoveSuffix {
        variable: String,
        pattern: Box<BashExpr>,
    },
    RemovePrefix {
        variable: String,
        pattern: Box<BashExpr>,
    },
    RemoveLongestPrefix {
        variable: String,
        pattern: Box<BashExpr>,
    },
    RemoveLongestSuffix {
        variable: String,
        pattern: Box<BashExpr>,
    },

}

Expression-level AST node

Variants

Literal(String)

String literal

Variable(String)

Variable reference: $VAR or ${VAR}

CommandSubst(Box<BashStmt>)

Command substitution: $(cmd) or cmd

Arithmetic(Box<ArithExpr>)

Arithmetic expansion: $((expr))

Array(Vec<BashExpr>)

Array/list: (item1 item2 item3)

Concat(Vec<BashExpr>)

String concatenation

Test(Box<TestExpr>)

Test expression: [ expr ]

Glob(String)

Glob pattern: *.txt

DefaultValue

Default value expansion: ${VAR:-default} If variable is unset or null, use default value

Fields

variable: String

default: Box<BashExpr>

AssignDefault

Assign default value expansion: ${VAR:=default} If variable is unset or null, assign default value to variable and use it

Fields

variable: String

default: Box<BashExpr>

ErrorIfUnset

Error if unset expansion: ${VAR:?message} If variable is unset or null, exit with error message

Fields

variable: String

message: Box<BashExpr>

AlternativeValue

Alternative value expansion: ${VAR:+alt_value} If variable is set and non-null, use alt_value, otherwise empty string

Fields

variable: String

alternative: Box<BashExpr>

StringLength

String length expansion: ${#VAR} Get the length of the string value of variable

Fields

variable: String

RemoveSuffix

Remove suffix expansion: ${VAR%pattern} Remove shortest matching suffix pattern from variable

Fields

variable: String

pattern: Box<BashExpr>

RemovePrefix

Remove prefix expansion: ${VAR#pattern} Remove shortest matching prefix pattern from variable

Fields

variable: String

pattern: Box<BashExpr>

RemoveLongestPrefix

Remove longest prefix expansion: ${VAR##pattern} Remove longest matching prefix pattern from variable (greedy)

Fields

variable: String

pattern: Box<BashExpr>

RemoveLongestSuffix

Remove longest suffix expansion: ${VAR%%pattern} Remove longest matching suffix pattern from variable (greedy)

Fields

variable: String

pattern: Box<BashExpr>