bashrs::formal

Module abstract_state

Abstract state representation for formal verification

This module defines the abstract machine state used to formally specify the semantics of both rash AST and POSIX shell commands.

Structs

AbstractState

Abstract representation of the system state

Enums

FileSystemEntry

Entry in the abstract filesystem