bashrs

Module stdlib

This module provides support for stdlib functions that are transpiled to POSIX shell runtime functions.

Structs

StdlibFunction

Stdlib function metadata

Constants

STDLIB_FUNCTIONS

All stdlib functions

Functions

contains_injection_attempt 🔒

get_shell_function_name

Get the shell function name for a stdlib function

is_safe_default_value 🔒

is_stdlib_function

Check if a function name is a stdlib function

is_valid_var_name 🔒